SRC/pyproject.toml
[project]
name = "fehres"
version = "0.1.0"
description = "A Retrieval-Augmented Generation (RAG) system for document Q&A"
readme = "README.md"
requires-python = ">=3.11.14"
dependencies = [
    "aiofiles==23.2.1",
    "alembic==1.14.0",
    "asyncpg==0.30.0",
    "cohere==5.5.8",
    "docx2txt",
    "fastapi>=0.124.1",
    "google-genai",
    "langchain",
    "langchain-ollama",
    "openai>=1.75.0",
    "pgvector==0.4.0",
    "prometheus-client==0.21.1",
    "psycopg2-binary==2.9.10",
    "pydantic-settings>=2.4.0",
    "pymupdf==1.24.3",
    "python-dotenv==1.0.1",
    "python-multipart==0.0.9",
    "qdrant-client==1.10.1",
    "sqlalchemy==2.0.36",
    "tqdm",
    "uvicorn[standard]>=0.34.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "langchain-community>=0.0.20",
    "lxml>=4.9.0",
    "numpy",
    "rank-bm25",
    "joblib",
    "nltk",
    "playwright>=1.40.0",
    "huggingface-hub>=1.2.3",
    "langchain-huggingface",
    "sentence-transformers",
    "torch>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt==4.0.1",
    "PyJWT>=2.8.0",
    "slowapi>=0.1.9",
    "email-validator>=2.0.0",
    "httpx>=0.25.0",
    "google-adk==1.31.0",
    "litellm>=1.75.5",
]
[project.optional-dependencies]
# Kept for backwards compatibility; playwright is now a main dependency
scraping-browser = ["playwright>=1.40.0"]

[tool.uv]
link-mode = "copy"

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
